# openapi.json
{
  "components": {
    "schemas": {
...
      "CancellationRequest": {
        "description": "Optional metadata accompanying a `POST /jobs/{id}/cancel` request body.\n\nAll fields are optional so existing clients (empty POST body) continue\nworking. The route handler fills missing fields with sensible defaults:\n`source = CancelSource::Api`, `host = <peer addr>`, others left null.\n\n**Why optional rather than required:** the same endpoint is hit by the\nTUI (which knows everything), the dashboard (which knows source + host),\nscripted curl (which knows nothing), and the staging orchestrator\n(which forwards a cancel from another server). Forcing required fields\nwould break all but the TUI; the route handler does enrichment instead.",
        "properties": {
          "correlation_id": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CorrelationId",
                "description": "Caller-side correlation ID for cross-system tracing."
              },
              {
                "type": "null"
              }
            ]
          },
          "host": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CallerHost",
                "description": "Caller's hostname or IP. Server fills with peer-addr if unspecified."
              },
              {
                "type": "null"
              }
            ]
          },
          "in_flight_filename": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DisplayPath",
                "description": "Filename being processed at cancel time (TUI snapshots from\n`runtime.state` so we can show \"cancel pending; finishing X\" in\nthe cancel-feedback render)."
              },
              {
                "type": "null"
              }
            ]
          },
          "pid": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CallerPid",
                "description": "Caller process ID, if known."
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CancelReason",
                "description": "Free-form reason (\"user-pressed-cancel\", \"ctrl-c-shutdown\", etc.)."
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CancelSource",
                "description": "Where the cancel originated. Defaults to `Api` if unspecified."
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Standard error response body, matching FastAPI's `HTTPException` format.\n\nAll non-2xx responses from the server use this shape so that clients\ncan parse errors uniformly. The one exception is\n[`ServerError::JobConflict`](crate::error::ServerError::JobConflict),\nwhich nests a `conflicts` array inside `detail`.",
        "properties": {
          "detail": {
            "description": "Human-readable description of what went wrong. For validation\nerrors this includes the specific field or constraint that failed.",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "JobId": {
        "description": "Server-assigned identifier for a job (non-empty).",
        "type": "string"
      },
...
      "JobListItem": {
        "description": "Summary for job listing (`GET /jobs` response element).\n\nA lighter-weight projection of [`JobInfo`] without per-file statuses,\nsuitable for listing many jobs at once.",
        "properties": {
          "active_lease": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LeaseRecord",
                "description": "Active lease information when this job is currently claimed by a node."
              },
              {
                "type": "null"
              }
            ]
          },
          "command": {
            "$ref": "#/components/schemas/ReleasedCommand",
            "description": "Batchalign command (e.g. \"morphotag\", \"align\")."
          },
          "completed_at": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MachineTime",
                "description": "When the job reached a terminal state (`MachineTime`)."
              },
              {
                "type": "null"
              }
            ]
          },
          "completed_files": {
            "description": "Number of files that finished successfully (`Done`).",
            "format": "int64",
            "type": "integer"
          },
          "control_plane": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JobControlPlaneInfo",
                "description": "Server control-plane metadata for this job when returned by a server."
              },
              {
                "type": "null"
              }
            ]
          },
          "duration_s": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/NonNegativeSeconds",
                "description": "Wall-clock duration in seconds.  `None` while the job is active."
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "description": "Job-level error message when the job failed (the aggregated per-file\nfailure reason). Surfaces the cause in the `/jobs` list and the live\ndashboard/TUI feed instead of a bare \"failed\". `None` for non-failed\njobs. Distinct from per-file errors and from the `error_files` count.",
            "type": [
              "string",
              "null"
            ]
          },
          "error_files": {
            "description": "Number of files that ended in `Error` status.",
            "format": "int64",
            "type": "integer"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Server-assigned UUID (v4) for this job."
          },
          "lang": {
            "$ref": "#/components/schemas/LanguageSpec",
            "description": "Job language: a resolved ISO 639-3 code, `\"auto\"`, `\"per-file\"`, or a\ncode-switched pair such as `\"eng,spa\"`."
          },
          "next_eligible_at": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MachineTime",
                "description": "When the job is deferred before execution, the earliest time\nat which it should be retried. `None` when immediately eligible."
              },
              {
                "type": "null"
              }
            ]
          },
          "num_workers": {
            "description": "Number of concurrent Python workers used for this job.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "source_dir": {
            "description": "Client's original input directory path, for display.",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current lifecycle state of the job."
          },
          "submitted_at": {
            "$ref": "#/components/schemas/MachineTime",
            "description": "When the job was submitted (`MachineTime`)."
          },
          "submitted_by": {
            "description": "IP address or Tailscale identifier of the submitting client.",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by_name": {
            "description": "Human-readable hostname of the submitting machine.",
            "type": [
              "string",
              "null"
            ]
          },
          "total_files": {
            "description": "Total number of files in this job.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "status",
          "command",
          "total_files",
          "completed_files",
          "submitted_at"
        ],
        "type": "object"
      },
...
      "JobSubmission": {
        "description": "`POST /jobs` request body.",
        "properties": {
          "before_paths": {
            "description": "Absolute paths to \"before\" files for incremental processing\n(paths_mode only). When non-empty, the diff engine compares each\nbefore file against its corresponding source_path and only\nreprocesses changed utterances.\n\nMust be the same length as `source_paths` when non-empty.",
            "items": {
              "$ref": "#/components/schemas/ClientPath"
            },
            "type": "array"
          },
          "command": {
            "$ref": "#/components/schemas/ReleasedCommand",
            "description": "Batchalign command (align, morphotag, etc.)."
          },
          "debug_traces": {
            "description": "When true, the server collects detailed algorithm traces for\nvisualization (DP alignment matrices, ASR pipeline stages, FA\ntimelines, retokenization mappings). Defaults to false, zero\noverhead when off.",
            "type": "boolean"
          },
          "display_names": {
            "description": "Human-readable filenames for display (paths_mode only, optional).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "files": {
            "description": "CHAT files to process.",
            "items": {
              "$ref": "#/components/schemas/FilePayload"
            },
            "type": "array"
          },
          "lang": {
            "$ref": "#/components/schemas/LanguageSpec",
            "description": "Job language: a 3-letter ISO code, `\"auto\"` for ASR-driven detection,\n`\"per-file\"`, or a code-switched pair such as `\"eng,spa\"`."
          },
          "media_files": {
            "description": "Media filenames for the server to resolve from media_roots (transcribe only).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "media_mapping": {
            "$ref": "#/components/schemas/MediaMappingKey",
            "description": "Key into server's media_mappings config (e.g. \"childes-data\")."
          },
          "media_subdir": {
            "$ref": "#/components/schemas/RepoRelativePath",
            "description": "Subdirectory under the mapped root (e.g. \"Eng-NA/MacWhinney/0young-ASR\")."
          },
          "num_speakers": {
            "$ref": "#/components/schemas/NumSpeakers",
            "description": "Number of speakers."
          },
          "options": {
            "description": "Typed command options (engine selections, processing flags, etc.)."
          },
          "output_paths": {
            "description": "Absolute paths to write output files to (paths_mode only).",
            "items": {
              "$ref": "#/components/schemas/ClientPath"
            },
            "type": "array"
          },
          "paths_mode": {
            "description": "When true, server reads/writes files directly via source_paths/output_paths.",
            "type": "boolean"
          },
          "source_dir": {
            "$ref": "#/components/schemas/ClientPath",
            "description": "Client's input directory path (for dashboard display)."
          },
          "source_paths": {
            "description": "Absolute paths to read input files from (paths_mode only).",
            "items": {
              "$ref": "#/components/schemas/ClientPath"
            },
            "type": "array"
          }
        },
        "required": [
          "command",
          "options"
        ],
        "type": "object"
      },
...
      "MachineTime": {
        "description": "RFC 3339 in UTC with exactly three fractional digits, so string order is time order.",
        "examples": [
          "2026-09-28T16:00:00.000Z"
        ],
        "format": "date-time",
        "type": "string"
      },
...
      "NonNegativeSeconds": {
        "description": "A finite, non-negative number of seconds.",
        "format": "double",
        "minimum": 0,
        "type": "number"
      },
...
      "StatusMessageResponse": {
        "description": "Response body for mutating operations that return a status confirmation\n(e.g. cancel, delete, restart).\n\nReturned with HTTP 200 on success. Clients can branch on `status` for\nprogrammatic handling and display `message` to users.",
        "properties": {
          "message": {
            "description": "Human-readable sentence describing the action taken (e.g.\n`\"Job abc123 has been cancelled.\"`).",
            "type": "string"
          },
          "status": {
            "description": "Machine-readable outcome token (e.g. `\"cancelled\"`, `\"deleted\"`,\n`\"restarted\"`). Suitable for switch/match in client code.",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "batchalign3: command-line interface, application, server, worker pool, cache",
    "license": {
      "identifier": "BSD-3-Clause",
      "name": "BSD-3-Clause"
    },
    "title": "batchalign",
    "version": "0.8.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "get": {
        "description": "Used by the dashboard and CLI `jobs` command. The response is intentionally\ncompact -- per-file detail is omitted and available via `GET /jobs/{id}`.",
        "operationId": "list_jobs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobListItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List all jobs"
          }
        },
        "summary": "Return a summary of every job the server knows about (active, completed,\nfailed, or cancelled).",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Validates the command against built-in tasks and worker-advertised capabilities,\nstages input files (content mode) or records source paths (paths mode), detects\n`(submitted_by, filename)` conflicts with active jobs, and spawns a background\nrunner task that acquires workers and dispatches files. The response echoes the\njob back as `JobInfo` and includes the correlation ID in `X-Request-Id`.",
        "operationId": "submit_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Submitted job"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Accept a new processing job and begin execution.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}": {
      "delete": {
        "description": "Returns 409 if the job is still running -- the caller must cancel it first.\nDeleting a job removes it from the in-memory store and SQLite, and broadcasts\na `JobDeleted` event to connected WebSocket/SSE clients.",
        "operationId": "delete_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessageResponse"
                }
              }
            },
            "description": "Job deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job still running"
          }
        },
        "summary": "Permanently remove a terminal job and its associated state.",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "This is the primary polling endpoint for the CLI: the client calls it\nrepeatedly until `status` reaches a terminal value.",
        "operationId": "get_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Job detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Return full detail for a single job, including per-file statuses.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "description": "Fires the job's `CancellationToken`, which the runner checks between files.\nAlready-terminal jobs (completed, failed, cancelled) are no-ops at the\nstate level, but the audit row is still recorded with `accepted=false`\nso a forensic reader sees every cancel gesture (e.g., a user pressing\ncancel twice an hour apart against a job that was already finishing).\nInterrupted jobs can still be cancelled -- they are not considered\nterminal here because they may be holding worker resources.\n\n**Body** is optional. When omitted, defaults to `source=Api` with the\npeer address as `host`. When present, fields override those defaults\nso the TUI can attribute its cancels to `source=Tui` and identify the\nin-flight file at the moment of the cancel.",
        "operationId": "cancel_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancellationRequest"
              }
            }
          },
          "description": "Optional caller provenance",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessageResponse"
                }
              }
            },
            "description": "Cancel request accepted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Request cancellation of a running or queued job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/restart": {
      "post": {
        "description": "Only jobs in a terminal non-cancelled state can be restarted. The store\nresets per-file statuses and clears errors, then a fresh runner task is\nspawned. This is the primary recovery path after transient worker crashes\nor OOM kills.",
        "operationId": "restart_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Job restarted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not restartable"
          }
        },
        "summary": "Reset a failed or interrupted job back to `Queued` and re-run it.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/results": {
      "get": {
        "description": "Blocks until the job has completed, failed, or been interrupted -- returns\n409 for jobs that are still running or queued. The response includes the\nfull staged content of each successful output file. Paths-mode jobs keep\nthe execution-host write, but also preserve a staged copy for download.\nCancelled jobs are intentionally excluded because they may have incomplete\nresults.",
        "operationId": "get_results",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultResponse"
                }
              }
            },
            "description": "Results for terminal jobs"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Retrieve all output files for a terminal job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/results/{*filename}": {
      "get": {
        "description": "Waits for the individual file to reach a terminal status, not the whole job,\nso clients can fetch results incrementally as files finish. Supports\ncommand-policy matching (e.g., requesting `sample.wav` finds `sample.cha`,\nand speaker identification maps `sample.cha` to its evidence JSON).",
        "operationId": "get_single_result",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Requested output filename",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResult"
                }
              }
            },
            "description": "One file result"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "File not found"
          }
        },
        "summary": "Retrieve the output for a single file within a job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/stream": {
      "get": {
        "description": "Sends:\n- `snapshot` event with current file statuses on connect\n- `file_update` events as files are processed\n- `complete` event when the job reaches a terminal state, then closes",
        "operationId": "stream_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE event stream"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "SSE stream for real-time per-file progress of a single job.",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}